requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
boto3==1.42.63
botocore==1.42.63
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0.1
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
docker==7.1.0
filelock==3.20.3
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
kombu==5.6.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pi==0.1.2
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydicom==3.0.1
python-dateutil==2.9.0.post0
python-magic
redis==7.1.1
requests==2.32.5
s3transfer==0.16.0
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
